sync help (shane)
Shane here is where I'm stuck at in dealing with syncing the client and server to reduce lag.

You're going to need to dump db as i've modified speed of characters down to an easier to work value (10 as opposed to 40).

Test moving just right.  In console of both node and client I'm echoing the player's x position whenever moving right is command is issued.  The client is updating much much faster (because update() on client is  natively using requestanimationframe rather than setInterval like the server is).  I tried to put a lock on how often update() on the client is being ran with a while statement, but its not keeping in sync.  need some help and may have to revert this commit, but I wanted to key in the lag trouble im having.